set player click

+++ include/gui/Gui.hpp
@@ -10,9 +10,9 @@
 #define GUI_TIME_LOOP 1000 / GUI_RATE
 
 #define GUI_WIN_W 800
-#define GUI_WIN_H (int)GUI_WIN_W * 0.7
+#define GUI_WIN_H (GUI_WIN_W * 0.7)
 #define GUI_BOARD_SZ GUI_WIN_H
-#define GUI_BOARD_START_X GUI_WIN_W - GUI_BOARD_SZ
+#define GUI_BOARD_START_X (int)(GUI_WIN_W - GUI_BOARD_SZ)
 
 // all colors are in format RGBA
 #define GUI_COLOR_1   0x000000FF
